refactor: [#1275] move conenction cookie to udp_tracker_core package
