+++ src/packages/udp_tracker_core/connection_cookie.rs
@@ -79,11 +79,24 @@
 
 use aquatic_udp_protocol::ConnectionId as Cookie;
 use cookie_builder::{assemble, decode, disassemble, encode};
+use thiserror::Error;
 use tracing::instrument;
 use zerocopy::AsBytes;
 
-use crate::servers::udp::error::ConnectionCookieError;
 use crate::shared::crypto::keys::CipherArrayBlowfish;
+
+/// Error returned when there was an error with the connection cookie.
+#[derive(Error, Debug)]
+pub enum ConnectionCookieError {
+    #[error("cookie value is not normal: {not_normal_value}")]
+    ValueNotNormal { not_normal_value: f64 },
+
+    #[error("cookie value is expired: {expired_value}, expected > {min_value}")]
+    ValueExpired { expired_value: f64, min_value: f64 },
+
+    #[error("cookie value is from future: {future_value}, expected < {max_value}")]
+    ValueFromFuture { future_value: f64, max_value: f64 },
+}
 
 /// Generates a new connection cookie.
 ///
