+++ src/servers/udp/error.rs
@@ -6,6 +6,8 @@ use derive_more::derive::Display;
 use thiserror::Error;
 use torrust_tracker_located_error::LocatedError;
 
+use crate::packages::udp_tracker_core::connection_cookie::ConnectionCookieError;
+
 #[derive(Display, Debug)]
 #[display(":?")]
 pub struct ConnectionCookie(pub ConnectionId);
@@ -42,19 +44,6 @@ pub enum Error {
     /// Error returned when tracker requires authentication.
     #[error("domain tracker requires authentication but is not supported in current UDP implementation. Location: {location}")]
     TrackerAuthenticationRequired { location: &'static Location<'static> },
-}
-
-/// Error returned when there was an error with the connection cookie.
-#[derive(Error, Debug)]
-pub enum ConnectionCookieError {
-    #[error("cookie value is not normal: {not_normal_value}")]
-    ValueNotNormal { not_normal_value: f64 },
-
-    #[error("cookie value is expired: {expired_value}, expected > {min_value}")]
-    ValueExpired { expired_value: f64, min_value: f64 },
-
-    #[error("cookie value is from future: {future_value}, expected < {max_value}")]
-    ValueFromFuture { future_value: f64, max_value: f64 },
 }
 
 impl From<RequestParseError> for Error {
